Move `SerializableClassInstance` to `src/serializable` and optimize imports
