+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from './ModifySourcePlugin';
 export * from './operations';
+export * from './range';
