+++ src/range/index.ts
@@ -1,0 +1,4 @@
+export * from './TextRange';
+export * from './LinesRange';
+export * from './MarkedTextRange';
+export * from './MarkedLinesRange';
