+++ src/serializable/SerializableClassInstance.ts
