+++ src/serializable/index.ts
@@ -1,0 +1,4 @@
+export * from './SerializableClassInstance';
+export * from './isSerializableOfClass';
+export * from './rangeFromSerializable';
+export * from './operationFromSerializable';
